Add vblast function to handle BLAST command
The function `vblast` has been introduced in `VERB.C` to execute the BLAST command. It checks the state of the second rod and whether the cave is closed, delivering a special bonus message and ending the game if conditions are met. The function operates on global game state variables and may print messages to the player. Previous related code has been removed.