Remove `idx4` array definition from ADVTEXT.H
The `idx4` array, previously defined as a `long` type with a size of `MAXMSG` and initialized with a series of values, has been removed from the `ADVTEXT.H` file. This change was made as the array is no longer required, potentially due to refactoring, optimization, or changes in program requirements.